asReadOnlyBuffer non native return
This may seem far-fetched, and the test cases to prove it is painful and
extremely difficult to reproduce unless you have the particular setup
It has only occurred on one setup a Samsung Galaxy Tab A 10.1 after OS
upgrade to Android sdk 24 (Android N/7.0)
However the bug was present consistently and the system could not run
Java3D
 
The returned read-only buffer should be a view of the original, sharing
data and byte order however the OS returned a copy of the data with a
non-native byte order.
Suffice to say the read buffer in that case is useless and almost
guaranteed fatal as it hands out garbage data to any requests.
The lack of mutability and introspection allowed by the read-only buffer
means the issue can't be discovered nor remedied at run time.