Maven publishing configuration setup
This change has removed older jaring scripts, and some distribution
libraries in favor of Maven creating a fully working deployed jar. Note
that Android Studio restricts duplicate classes from being in a project
at all, which forces java3d-core-and to merge it's target jar with
java3-core as well as suppress publication of the non "-android" version
of the jogl libraries.
Finally there is an issue in the linux shared library dependencies
resolution in jogl so I have to include them in the target jar, this
will be reversed later.

Signed-off-by: Philip Jordan <p.j.nz@outlook.com>